1) обновил либу нотификаций 2) добавил возможность автообновлений

+++ main.py
@@ -9,34 +9,30 @@ import multiprocessing as mp
 from src.const import const
 from src.service.api import WtApi
 from src.service.presence import presence_service
-from src.service.win_notify import notify_service
+from src.service.win_notify import WinNotificationService
 from src.settings import settings
 
 
 async def main() -> None:
     await settings.set_settings()
+    await settings.store_sys_data(
+        pid=os.getpid(), os_data=settings.win_version,
+    )
     logger.info(f'App version -> {const.APP_VERSION}')
-    logger.info('Starting notify process')
     try:
-        if notify_service.win_version == '11':
-            logger.info('Start plyer loop')
+        notify_service = WinNotificationService(settings)
+        notify_service.start_notify(app_version=const.APP_VERSION)
+        if settings.show_update_notifications is True:
+            logger.info('Start notify loop')
             notify_process = mp.Process(
-                target=notify_service.win11_notify_loop,
-                args=(settings,),
+                target=notify_service.win_notify_loop,
                 name='NotifyProcess',
                 daemon=False,
             )
-        else:
-            logger.info('Start win10-toast loop')
-            notify_process = mp.Process(
-                target=notify_service.win10_notify_loop,
-                args=(settings,),
-                name='NotifyProcess',
-                daemon=False,
-            )
-        notify_process.start()
-        notify_pid = notify_process.pid
-        logger.info(f'Notify process pid: {notify_pid}')
+            notify_process.start()
+            notify_pid = notify_process.pid
+            await settings.add_notify_pid_to_sys_info(notify_pid)
+            logger.info(f'Notify process pid: {notify_pid}')
         while True:
             discord_status = await presence_service.get_discord_pipe()
             if discord_status is None:
@@ -65,14 +61,17 @@ async def main() -> None:
                     logger.warning(exc)
                     await presence_service.connect()
     finally:
-        logger.debug(f'Stopping notify process')
-        notify_process.terminate()
-        try:
-            os.kill(notify_pid, signal.SIGTERM)
-        except Exception as e:
-            logger.warning(e)
+        if settings.show_update_notifications is True:
+            logger.debug(f'Stopping notify process')
+            notify_process.terminate()
+            try:
+                os.kill(notify_pid, signal.SIGTERM)
+            except Exception as e:
+                logger.warning(e)
 
 if __name__ == '__main__':
     mp.freeze_support()
     mp.set_start_method('spawn', force=True)
     asyncio.run(main())
+    # updater = UpdaterService()
+    # updater.download_release_zip()
